requirements.txt
jupyter
boto3
streamlit
python-dotenv